refactor(server): relocate synchronous delegation into delegation.go
Move executeDelegatedTask, the workshop sub-agent tracking notifier, the
delegation start/end UI events, and their two types (~1,020 lines) out of
server.go into a new delegation.go. Async/background delegation already
moved to background_agents.go in the prior commit. Pure relocation: same
package, no behavior change.

server.go 10,189 → 9,167 lines (−2,571 from the original 11,738 across
both relocations). fmt/vet clean, cmd/server tests pass.
