+++ agent_go/cmd/server/server.go
@@ -54,7 +54,6 @@ import (
 	"mcp-agent-builder-go/agent_go/pkg/common"
 	"mcp-agent-builder-go/agent_go/pkg/logger"
 	"mcp-agent-builder-go/agent_go/pkg/skills"
-	"mcp-agent-builder-go/agent_go/pkg/subagents"
 
 	"github.com/joho/godotenv"
 
@@ -6639,1027 +6638,6 @@ func (e *sessionEventEmitter) EmitBlockingHumanFeedback(requestID, question, con
 	}
 	e.eventStore.AddEvent(e.sessionID, event)
 	log.Printf("[PLAN APPROVAL] Emitted blocking_human_feedback event for plan approval (request_id: %s, session: %s)", requestID, e.sessionID)
-}
-
-// executeDelegatedTask executes a delegated task via a sub-agent.
-// onCreated is an optional callback invoked after the sub-agent wrapper is created
-// but before Invoke — used by background agents to attach a history func.
-func (api *StreamingAPI) executeDelegatedTask(ctx context.Context, parentReq QueryRequest, sessionID string, instruction string, onCreated ...func(wrapper *agent.LLMAgentWrapper)) (string, error) {
-	log.Printf("[DELEGATION] Creating sub-agent for delegated task in session %s", sessionID)
-
-	// The full delegation contract (depth, tier, template, servers, skills,
-	// browser sharing, background link) arrives as one typed spec.
-	spec := virtualtools.SubAgentSpecFromContext(ctx)
-	currentDepth := spec.Depth
-
-	if currentDepth >= virtualtools.MaxDelegationDepth {
-		return "", fmt.Errorf("maximum delegation depth (%d) reached", virtualtools.MaxDelegationDepth)
-	}
-
-	// Generate a unique delegation ID for tracking
-	delegationID := fmt.Sprintf("delegation-%d-%d", currentDepth, time.Now().UnixNano())
-
-	// When spawned inside a workshop step execution, record the step→delegation mapping so
-	// query_step can surface tool calls made by this API-based sub-agent.
-	if forcedID, ok := ctx.Value(orchEvents.ForceCorrelationIDKey).(string); ok && strings.HasPrefix(forcedID, "workshop-step-") {
-		registerStepDelegation(forcedID, delegationID)
-	}
-
-	// Build sub-agent config from parent request
-	// Get provider and model from parent request
-	provider := llm.Provider(parentReq.Provider)
-	if provider == "" {
-		provider = llm.Provider("anthropic")
-	}
-	modelID := parentReq.ModelID
-	if modelID == "" {
-		modelID = "claude-sonnet-4-20250514"
-	}
-
-	// Load sub-agent template if specified
-	var loadedTemplate *subagents.SubAgent
-	agentTemplateName := spec.AgentTemplate
-	if agentTemplateName != "" {
-		workspaceAPIURL := getWorkspaceAPIURL()
-		sa, err := subagents.GetSubAgent(workspaceAPIURL, agentTemplateName)
-		if err != nil {
-			log.Printf("[DELEGATION] Warning: Failed to load sub-agent template %s: %v", agentTemplateName, err)
-		} else {
-			loadedTemplate = sa
-			log.Printf("[DELEGATION] Loaded sub-agent template: %s (%s)", sa.Frontmatter.Name, agentTemplateName)
-		}
-	}
-
-	// Resolve reasoning level tier to specific provider/model if configured
-	reasoningLevel := spec.ReasoningLevel
-	// Apply template defaults if not explicitly set
-	if reasoningLevel == "" && loadedTemplate != nil && loadedTemplate.Frontmatter.DefaultReasoningLevel != "" {
-		reasoningLevel = loadedTemplate.Frontmatter.DefaultReasoningLevel
-		log.Printf("[DELEGATION] Using template default reasoning_level: %s", reasoningLevel)
-	}
-	var tierFallbacks []agent.FallbackModel
-	if reasoningLevel != "" {
-		// Load fresh from workspace file at delegation time so LLM-written tier changes take effect immediately
-		tierConfig := LoadAndResolveTierConfig(ctx, parentReq.DelegationTierConfig)
-		if tierConfig != nil {
-			var tierModel *virtualtools.TierModel
-			switch reasoningLevel {
-			case "high":
-				tierModel = tierConfig.High
-			case "medium":
-				tierModel = tierConfig.Medium
-			case "low":
-				tierModel = tierConfig.Low
-			default:
-				// Custom tier lookup
-				if tierConfig.Custom != nil {
-					if ct, ok := tierConfig.Custom[reasoningLevel]; ok {
-						tierModel = &virtualtools.TierModel{Provider: ct.Provider, ModelID: ct.ModelID}
-					}
-				}
-			}
-			if tierModel != nil && tierModel.Provider != "" && tierModel.ModelID != "" {
-				provider = llm.Provider(tierModel.Provider)
-				modelID = tierModel.ModelID
-				tierFallbacks = convertTierFallbacksToAgentFallbacks(tierModel.Fallbacks, tierModel.Provider)
-				log.Printf("[DELEGATION] Using tier %s model: %s/%s", reasoningLevel, tierModel.Provider, tierModel.ModelID)
-			}
-		}
-	}
-
-	// Build server name — use delegation-specific servers if provided, otherwise all parent servers
-	var serverName string
-	var serversList []string
-	if len(spec.Servers) > 0 {
-		serverName = strings.Join(spec.Servers, ",")
-		serversList = spec.Servers
-		log.Printf("[DELEGATION] Using sub-agent specific servers: %s", serverName)
-	} else if len(parentReq.EnabledServers) > 0 {
-		serverName = strings.Join(parentReq.EnabledServers, ",")
-		serversList = parentReq.EnabledServers
-	} else if len(parentReq.Servers) > 0 {
-		serverName = strings.Join(parentReq.Servers, ",")
-		serversList = parentReq.Servers
-	}
-
-	// Sub-agents always run in code_execution mode (Python harness calling MCP tools via HTTP API).
-	useCodeExec := true
-
-	// Extract background agent ID if this delegation was spawned by a background agent
-	backgroundAgentID := spec.BackgroundAgentID
-
-	// Emit delegation_start event (after model and server resolution so we can include all info)
-	api.emitDelegationStartEvent(sessionID, delegationID, currentDepth, instruction, reasoningLevel, modelID, serversList, backgroundAgentID, agentTemplateName)
-
-	// Load merged API keys (env + workspace)
-	apiKeys := MergedProviderAPIKeys(ctx)
-
-	// Get user ID from context for per-user OAuth token isolation
-	subAgentUserID := ""
-	if userID, ok := ctx.Value(common.UserIDKey).(string); ok {
-		subAgentUserID = userID
-	}
-	log.Printf("[USER_ID_DEBUGGING] Sub-agent: subAgentUserID=%q (from parent context UserIDKey)", subAgentUserID)
-
-	// Determine sub-agent session ID: isolated when share_browser=false, shared otherwise
-	subAgentSessionID := sessionID
-	if !spec.ShareBrowser {
-		subAgentSessionID = fmt.Sprintf("%s-isolated-%d", sessionID, time.Now().UnixNano())
-		log.Printf("[DELEGATION] Browser isolation: sub-agent gets new session ID %s (parent: %s)", subAgentSessionID, sessionID)
-	}
-
-	// Create sub-agent config based on parent request
-	subAgentConfig := agent.LLMAgentConfig{
-		Name:       fmt.Sprintf("sub-agent-depth-%d", currentDepth),
-		ServerName: serverName,
-		ConfigPath: api.mcpConfigPath,
-		Provider:   provider,
-		ModelID:    modelID,
-		Temperature: func() float64 {
-			if parentReq.Temperature > 0 {
-				return parentReq.Temperature
-			}
-			return 0.7
-		}(),
-		MaxTurns: func() int {
-			if parentReq.MaxTurns != 0 {
-				return parentReq.MaxTurns
-			}
-			return 500
-		}(),
-		ToolChoice:         "", // Empty — let the library decide; Azure/OpenAI reject tool_choice when no tools are present
-		StreamingChunkSize: 1,
-		// No Timeout set — sub-agent lifetime is controlled by the parent context.
-		// Sub-agent mode uses the resolved values (from delegate call, template default, or auto-enable).
-		UseCodeExecutionMode:  useCodeExec,
-		APIKeys:               apiKeys,
-		Fallbacks:             tierFallbacks,
-		SessionID:             subAgentSessionID, // Reuse parent session's MCP connections via registry, unless browser isolation requested
-		UserID:                subAgentUserID,    // Per-user OAuth token isolation
-		CodingAgentWorkingDir: codingAgentWorkspaceWorkingDir(perUserChatsFolderFor(subAgentUserID)),
-	}
-	// Tool timeout, context summarization/editing, large-output offloading, and
-	// parallel tool execution inherit from the parent request the same way the
-	// root chat agent resolves them (no preset at delegation time).
-	applySharedLLMAgentTuning(&subAgentConfig, &parentReq, nil)
-
-	// Create sub-agent using the wrapper (same as parent agent creation)
-	subAgent, err := agent.NewLLMAgentWrapper(ctx, subAgentConfig, nil, api.logger)
-	if err != nil {
-		api.emitDelegationEndEvent(sessionID, delegationID, currentDepth, "", err.Error(), nil)
-		return "", fmt.Errorf("failed to create sub-agent: %w", err)
-	}
-
-	// Resolve conditional folder-guard grants for the sub-agent once.
-	// Used by both nested scopes below (prompt assembly + workspace tool folder guard).
-	subResolvedGrants := resolveConditionalGrants(parentReq)
-	browserReq := api.withEffectiveBrowserMode(ctx, parentReq, sessionID)
-	subBrowserCfg := buildChatBrowserConfig(browserReq)
-
-	// Add event observers to sub-agent so its events appear in the UI and
-	// its token usage lands in the global cost ledger.
-	if underlyingAgent := subAgent.GetUnderlyingAgent(); underlyingAgent != nil {
-		subAgentObserver := events.NewDelegationEventObserver(api.eventStore, sessionID, currentDepth, delegationID, api.logger)
-		if toolCb, ok := ctx.Value(virtualtools.ToolEventCallbackKey).(events.ToolEventCallback); ok && toolCb != nil {
-			subAgentObserver.OnToolEvent = toolCb
-		}
-		underlyingAgent.AddEventListener(subAgentObserver)
-		parentUserID, _ := ctx.Value(common.UserIDKey).(string)
-		underlyingAgent.AddEventListener(newCostObserver(api.costLedger, sessionID, parentUserID, parentReq.AgentMode))
-		log.Printf("[DELEGATION] Added event observers for sub-agent at depth %d", currentDepth)
-
-		// Phase 6 explicit-pass: sub-agents inherit NO skills from the
-		// parent. The parent must enumerate skills the sub-agent needs
-		// in its delegate() call (skills=[...]). delegation_tools.go
-		// threads those names through the SubAgentSpec.
-		if len(spec.Skills) > 0 {
-			if attached := skills.LoadAttachable(getWorkspaceAPIURL(), spec.Skills); len(attached) > 0 {
-				for _, s := range attached {
-					underlyingAgent.AttachSkill(s)
-				}
-				log.Printf("[DELEGATION] Attached %d skill(s) to sub-agent (explicit pass)", len(attached))
-			}
-		}
-
-		// Append prompt sections contributed by active conditional grants
-		// (resolved above in subResolvedGrants before this block).
-		for _, section := range subResolvedGrants.PromptSections {
-			underlyingAgent.AppendSystemPrompt(section)
-		}
-		if len(subResolvedGrants.PromptSections) > 0 {
-			log.Printf("[DELEGATION] Appended %d grant prompt section(s) to sub-agent: %v", len(subResolvedGrants.PromptSections), subResolvedGrants.AppliedNames)
-		}
-
-		// Inject sub-agent template instructions into system prompt
-		if loadedTemplate != nil {
-			templatePrompt := fmt.Sprintf("\n## Sub-Agent Role: %s\n\n%s\n",
-				loadedTemplate.Frontmatter.Name, loadedTemplate.Content)
-			underlyingAgent.AppendSystemPrompt(templatePrompt)
-			log.Printf("[DELEGATION] Injected sub-agent template instructions: %s", loadedTemplate.Frontmatter.Name)
-		}
-
-		// Merge global secrets with parent's decrypted secrets — inject names into prompt (values are in env vars)
-		allDelegationSecrets := mergeGlobalSecrets(parentReq.DecryptedSecrets, parentReq.SelectedGlobalSecrets)
-		if len(allDelegationSecrets) > 0 {
-			underlyingAgent.AppendSystemPrompt(buildSecretNamesPrompt(allDelegationSecrets))
-			log.Printf("[DELEGATION] Injected %d secret names (not values) into sub-agent system prompt", len(allDelegationSecrets))
-		}
-
-		// Give sub-agents the workspace folder structure so they know where to
-		// read/write files. Sub-agents are actual file workers that need this orientation.
-		// Use the same per-user Chats folder as the parent session.
-		subAgentChatsFolder := perUserChatsFolderFor(subAgentUserID)
-		subAgentMemoryFolder := perUserMemoryFolderFor(subAgentUserID)
-		subShellRoot := fsutil.WorkspaceShellRoot()
-		underlyingAgent.AppendSystemPrompt(GetWorkspaceMap(subShellRoot, subAgentChatsFolder, subAgentMemoryFolder))
-		underlyingAgent.AppendSystemPrompt(GetWorkspaceReference(subShellRoot, subAgentChatsFolder, subAgentMemoryFolder))
-		log.Printf("[DELEGATION] Added workspace instructions to sub-agent (chats=%s)", subAgentChatsFolder)
-
-		// Give sub-agents access to memory tools so they can persist key discoveries
-		// across tasks (reads from Chats/memories/ by default).
-		api.activeSessionsMux.RLock()
-		subAgentMemFolder := ""
-		if sess, ok := api.activeSessions[sessionID]; ok {
-			subAgentMemFolder = sess.MemoryFolder
-		}
-		api.activeSessionsMux.RUnlock()
-		underlyingAgent.AppendSystemPrompt(virtualtools.GetMemoryInstructions(subAgentMemFolder))
-		log.Printf("[DELEGATION] Added memory instructions to sub-agent")
-
-		// [BROWSER] Add browser instructions using standardized builder (same as parent chat agent).
-		// Sub-agents need their own transformer registration because each Agent instance has
-		// its own toolArgTransformers map — the parent's transformer doesn't propagate.
-		if subBrowserPrompt := browserinstructions.BuildBrowserInstructions(subBrowserCfg); subBrowserPrompt != "" {
-			underlyingAgent.AppendSystemPrompt(subBrowserPrompt)
-			log.Printf("[BROWSER] Added browser instructions to sub-agent (playwright=%v, agent-browser=%v, cdp=%v)",
-				subBrowserCfg.HasPlaywright, subBrowserCfg.HasAgentBrowser, subBrowserCfg.CdpPort > 0)
-		}
-
-		// Register file path transformer for browser file uploads on sub-agent
-		hasBrowserAccess := subBrowserCfg.HasAgentBrowser
-		hasPlaywright := subBrowserCfg.HasPlaywright
-		for _, s := range browserReq.EnabledServers {
-			if s == "playwright" {
-				hasPlaywright = true
-			}
-		}
-		if hasBrowserAccess || hasPlaywright {
-			wsAbsPath := fsutil.WorkspaceShellRoot()
-			underlyingAgent.SetToolArgTransformer("browser_file_upload", func(args map[string]interface{}) {
-				paths, ok := args["paths"].([]interface{})
-				if !ok || len(paths) == 0 {
-					log.Printf("[BROWSER_UPLOAD] Sub-agent: no paths in args, skipping transform")
-					return
-				}
-				for i, p := range paths {
-					pathStr, ok := p.(string)
-					if !ok || pathStr == "" || filepath.IsAbs(pathStr) {
-						continue
-					}
-					resolved := filepath.Join(wsAbsPath, pathStr)
-					log.Printf("[BROWSER_UPLOAD] Sub-agent resolved path[%d]: %q -> %q", i, pathStr, resolved)
-					paths[i] = resolved
-				}
-			})
-			log.Printf("[BROWSER_UPLOAD] Registered sub-agent browser_file_upload transformer, workspace=%s", wsAbsPath)
-		}
-
-		// Browser isolation: when share_browser=false, tell the sub-agent to use a unique
-		// session name with the agent_browser tool to avoid sharing browser state.
-		if !spec.ShareBrowser {
-			underlyingAgent.AppendSystemPrompt(fmt.Sprintf("## Browser Isolation\nYou have an isolated browser session. When using the agent_browser tool, use a unique session name (e.g., \"isolated-%d\") instead of \"default\" to avoid sharing browser state with other agents.", time.Now().UnixNano()))
-			log.Printf("[DELEGATION] Added browser isolation guidance to sub-agent system prompt")
-		}
-	}
-
-	// Register workspace tools for sub-agent
-	if underlyingAgent := subAgent.GetUnderlyingAgent(); underlyingAgent != nil {
-		// Sub-agents get the normal LLM-visible workspace tool set (advanced + media/provider tools).
-		workspaceRegistry := virtualtools.CreateWorkspaceToolRegistry(virtualtools.WorkspaceToolRegistryConfig{
-			WorkspaceAPIURL: getWorkspaceAPIURL(),
-			UserID:          subAgentUserID,
-			SessionID:       sessionID,
-		})
-		workspaceTools := workspaceRegistry.Tools
-		workspaceExecutors := workspaceRegistry.Executors
-		subAgentEnv := workspaceRegistry.Env
-		toolCategories := workspaceRegistry.Categories
-		log.Printf("[USER_ID_DEBUGGING] Sub-agent workspace executors: created with explicit userID=%q sessionID=%q", subAgentUserID, sessionID)
-		// Inject secrets as environment variables for sub-agent shell execution (SECRET_ prefix for whitelist)
-		delegationSecrets := mergeGlobalSecrets(parentReq.DecryptedSecrets, parentReq.SelectedGlobalSecrets)
-		if subAgentEnv != nil && len(delegationSecrets) > 0 {
-			for _, s := range delegationSecrets {
-				subAgentEnv["SECRET_"+s.Name] = s.Value
-			}
-			log.Printf("[SECRETS] Injected %d secrets as env vars for sub-agent shell execution", len(delegationSecrets))
-		}
-		// Inject LLM config fallback for read_image HTTP calls (e.g., from claude CLI subprocess)
-		if underlying := subAgent.GetUnderlyingAgent(); underlying != nil {
-			virtualtools.SetReadImageFallbackLLMConfig(workspaceExecutors, underlying.GetLLMModelConfig())
-		}
-
-		// Conditional grants already resolved above into subResolvedGrants.
-		// Merge parent @context paths and #workflow references into delegated folder-guard access.
-		// @context paths get write access; #workflow paths get read-only access.
-		fileContextWriteFolders, workflowReadOnlyFolders := collectSplitFolderGuardFolders(parentReq.Query, parentReq.WorkflowContextPaths)
-		if len(fileContextWriteFolders) > 0 {
-			log.Printf("[DELEGATION] Extracted write folder-guard paths from parent @context: %v", fileContextWriteFolders)
-		}
-		if len(workflowReadOnlyFolders) > 0 {
-			log.Printf("[DELEGATION] Extracted read-only folder-guard paths from parent #workflow: %v", workflowReadOnlyFolders)
-		}
-
-		// Apply per-user folder guard for all sub-agents.
-		// Writes scoped to _users/<subAgentUserID>/Chats/ and _users/<subAgentUserID>/memories/.
-		{
-			subPerUserChatsFolder := perUserChatsFolderFor(subAgentUserID)
-			subPerUserChatsWrite := subPerUserChatsFolder + "/"
-			subPerUserMemWrite := perUserMemoryFolderFor(subAgentUserID) + "/"
-			subPerUserChatHistory := strings.TrimSuffix(subPerUserChatsFolder, "Chats") + "chat_history/"
-			extraFolders := append([]string{}, subResolvedGrants.WriteFolders...)
-			extraFolders = append(extraFolders, fileContextWriteFolders...)
-			extraFolders = append(extraFolders, subPerUserMemWrite)
-			extraFolders = append(extraFolders, subPerUserChatHistory)
-			// Delegation path has no #workflow-derived blocked-write prefix (the parent
-			// session's blocked paths aren't inherited here; this call path is for sub-agents
-			// spawned with their own folder scope). Pass nil.
-			workspaceExecutors = wrapExecutorsWithChatModeFolderGuard(workspaceExecutors, workflowReadOnlyFolders, nil, extraFolders...)
-			workspace.SetSessionWorkingDir(sessionID, subPerUserChatsFolder)
-			readPaths := append([]string{subPerUserChatsWrite, subPerUserChatHistory, "Downloads/", "skills/", "subagents/", "Workflow/", subPerUserMemWrite}, extraFolders...)
-			readPaths = append(readPaths, subResolvedGrants.ReadOnlyExtra...)
-			readPaths = append(readPaths, workflowReadOnlyFolders...)
-			workspace.SetSessionFolderGuard(sessionID,
-				readPaths,
-				append([]string{subPerUserChatsWrite, "Downloads/", subPerUserMemWrite, subPerUserChatHistory}, extraFolders...),
-			)
-		}
-
-		// Register workspace tools
-		for _, tool := range workspaceTools {
-			if tool.Function == nil {
-				continue
-			}
-			toolName := tool.Function.Name
-			if executor, exists := workspaceExecutors[toolName]; exists {
-				enhancedDescription := enhanceToolDescriptionForChatMode(toolName, tool.Function.Description, perUserChatsFolderFor(subAgentUserID))
-
-				var params map[string]interface{}
-				if tool.Function.Parameters != nil {
-					paramsBytes, err := json.Marshal(tool.Function.Parameters)
-					if err == nil {
-						json.Unmarshal(paramsBytes, &params)
-					}
-				}
-				if params == nil {
-					continue
-				}
-
-				toolCategory := toolCategories[toolName]
-				if toolCategory == "" {
-					continue
-				}
-
-				if virtualtools.IsImageTool(toolName) && parentReq.ImageGenConfig != nil {
-					executor = virtualtools.WrapImageToolExecutorWithRuntimeOverride(executor, virtualtools.ImageGenRuntimeOverride{
-						Provider: parentReq.ImageGenConfig.Provider,
-						ModelID:  parentReq.ImageGenConfig.ModelID,
-						APIKey:   parentReq.ImageGenConfig.APIKey,
-					})
-				}
-
-				if err := underlyingAgent.RegisterCustomTool(
-					toolName,
-					enhancedDescription,
-					params,
-					executor,
-					toolCategory,
-				); err != nil {
-					log.Printf("[DELEGATION] Warning: Failed to register tool %s for sub-agent: %v", toolName, err)
-				}
-			}
-		}
-
-		// Register browser tools if enabled
-		if subBrowserCfg.HasAgentBrowser {
-			browserTools := virtualtools.CreateWorkspaceBrowserTools()
-			browserExecutors := virtualtools.CreateWorkspaceBrowserToolExecutorsWithSession(sessionID, getCdpPort(browserReq))
-			browserCategory := virtualtools.GetWorkspaceBrowserToolCategory()
-
-			browserExtraFolders := append([]string{}, subResolvedGrants.WriteFolders...)
-			browserExtraFolders = append(browserExtraFolders, fileContextWriteFolders...)
-			browserExecutors = wrapExecutorsWithChatModeFolderGuard(browserExecutors, workflowReadOnlyFolders, nil, browserExtraFolders...)
-
-			for _, tool := range browserTools {
-				if tool.Function == nil {
-					continue
-				}
-				toolName := tool.Function.Name
-				if executor, exists := browserExecutors[toolName]; exists {
-					var params map[string]interface{}
-					if tool.Function.Parameters != nil {
-						paramsBytes, err := json.Marshal(tool.Function.Parameters)
-						if err == nil {
-							json.Unmarshal(paramsBytes, &params)
-						}
-					}
-					if params == nil {
-						continue
-					}
-
-					if err := underlyingAgent.RegisterCustomTool(
-						toolName,
-						tool.Function.Description,
-						params,
-						executor,
-						browserCategory,
-					); err != nil {
-						log.Printf("[DELEGATION] Warning: Failed to register browser tool %s for sub-agent: %v", toolName, err)
-					}
-				}
-			}
-		}
-
-		// NOTE: Sub-agents do NOT get the delegate tool themselves (v1 design choice)
-		// This prevents runaway delegation chains.
-
-		// Minimal worker context — tells the sub-agent its role and output conventions.
-		subWorkerChatsFolder := perUserChatsFolderFor(subAgentUserID)
-		underlyingAgent.AppendSystemPrompt(fmt.Sprintf(`## Your Role
-You are a focused background worker. Complete the assigned task using available tools and return a clear, concise result.
-- You cannot spawn further sub-agents
-- You have no shared memory with the caller — all context is in the instruction you received
-- Save any output files under %s/ (use the sub-folder specified in your instruction, or create a descriptive one if none is given)
-- Return a summary of what you did and what you found`, subWorkerChatsFolder))
-		log.Printf("[DELEGATION] Added worker context to sub-agent (chats=%s)", subWorkerChatsFolder)
-	}
-
-	log.Printf("[DELEGATION] Sub-agent created, executing instruction at depth %d", currentDepth)
-
-	// Notify caller that the sub-agent wrapper is ready (used by background agents)
-	if len(onCreated) > 0 && onCreated[0] != nil {
-		onCreated[0](subAgent)
-	}
-
-	// Clean up isolated browser session when sub-agent finishes
-	if subAgentSessionID != sessionID {
-		defer func() {
-			mcpagent.CloseSession(subAgentSessionID)
-			log.Printf("[DELEGATION] Closed isolated browser session %s", subAgentSessionID)
-		}()
-	}
-
-	// Run the sub-agent with the instruction
-	startTime := time.Now()
-	result, err := subAgent.Invoke(ctx, instruction)
-	duration := time.Since(startTime)
-
-	// Collect metrics from sub-agent
-	metrics := subAgent.GetMetricsSnapshot()
-	stats := &delegationEndStats{
-		InputTokens:  metrics.InputTokens,
-		OutputTokens: metrics.OutputTokens,
-		ToolCalls:    metrics.ToolCallsExecuted,
-		Duration:     duration.String(),
-		TotalCostUSD: metrics.TotalCostUSD,
-	}
-
-	if err != nil {
-		api.emitDelegationEndEvent(sessionID, delegationID, currentDepth, "", err.Error(), stats)
-		return "", fmt.Errorf("sub-agent execution failed: %w", err)
-	}
-
-	// Emit delegation_end event with success
-	api.emitDelegationEndEvent(sessionID, delegationID, currentDepth, fmt.Sprintf("Completed in %s", duration), "", stats)
-
-	log.Printf("[DELEGATION] Sub-agent completed at depth %d in %s", currentDepth, duration)
-	return result, nil
-}
-
-// --- Background Agent Infrastructure for Async Delegation ---
-
-// bgAgentQuerierImpl implements virtualtools.BGAgentQuerier using the registry
-type bgAgentQuerierImpl struct {
-	registry *BackgroundAgentRegistry
-}
-
-func (q *bgAgentQuerierImpl) QueryAgent(sessionID, agentID string, last, offset int) (*virtualtools.BGAgentInfo, error) {
-	agent := q.registry.Get(sessionID, agentID)
-	if agent == nil {
-		return nil, fmt.Errorf("agent %s not found", agentID)
-	}
-	snap := agent.GetSnapshot()
-	elapsed := time.Since(snap.CreatedAt)
-	if snap.CompletedAt != nil {
-		elapsed = snap.CompletedAt.Sub(snap.CreatedAt)
-	}
-	info := &virtualtools.BGAgentInfo{
-		ID:        snap.ID,
-		Name:      snap.Name,
-		Status:    string(snap.Status),
-		Elapsed:   elapsed.Truncate(time.Second).String(),
-		CreatedAt: snap.CreatedAt.Format(time.RFC3339),
-	}
-	if snap.CompletedAt != nil {
-		info.CompletedAt = snap.CompletedAt.Format(time.RFC3339)
-	}
-	if snap.Status == BGAgentCompleted || snap.Status == BGAgentFailed {
-		info.Result = truncateForToolResponse(snap.Result, 4000)
-		info.Error = snap.Error
-	}
-	if snap.Status == BGAgentRunning {
-		// Return conversation history with pagination (last N entries, skip offset from end)
-		agent := q.registry.Get(sessionID, agentID)
-		if agent != nil {
-			// Get more entries than needed so we can apply offset
-			allHistory := agent.GetRecentHistory(last + offset)
-			// Apply offset: trim the last `offset` entries
-			if offset > 0 && len(allHistory) > offset {
-				allHistory = allHistory[:len(allHistory)-offset]
-			} else if offset > 0 {
-				allHistory = nil // offset exceeds history length
-			}
-			// Take only the last `last` entries
-			if len(allHistory) > last {
-				allHistory = allHistory[len(allHistory)-last:]
-			}
-			for _, h := range allHistory {
-				info.RecentHistory = append(info.RecentHistory, virtualtools.BGAgentHistoryEntry{
-					Role: h.Role,
-					Text: truncateForToolResponse(h.Text, 1000),
-				})
-			}
-		}
-		// Include recent tool calls with timing
-		if agent != nil {
-			toolCalls := agent.GetRecentToolCalls(5)
-			for _, tc := range toolCalls {
-				dur := ""
-				if tc.Status == "running" {
-					dur = time.Since(tc.StartedAt).Truncate(time.Second).String()
-				} else if tc.Duration > 0 {
-					dur = tc.Duration.Truncate(time.Millisecond).String()
-				}
-				info.RecentToolCalls = append(info.RecentToolCalls, virtualtools.BGAgentToolCall{
-					ToolName: tc.ToolName,
-					Duration: dur,
-					Status:   tc.Status,
-				})
-			}
-		}
-	}
-	return info, nil
-}
-
-func (q *bgAgentQuerierImpl) ListAgents(sessionID string) ([]*virtualtools.BGAgentInfo, error) {
-	agents := q.registry.GetAll(sessionID)
-	infos := make([]*virtualtools.BGAgentInfo, 0, len(agents))
-	for _, agent := range agents {
-		snap := agent.GetSnapshot()
-		elapsed := time.Since(snap.CreatedAt)
-		if snap.CompletedAt != nil {
-			elapsed = snap.CompletedAt.Sub(snap.CreatedAt)
-		}
-		info := &virtualtools.BGAgentInfo{
-			ID:      snap.ID,
-			Name:    snap.Name,
-			Status:  string(snap.Status),
-			Elapsed: elapsed.Truncate(time.Second).String(),
-		}
-		if snap.Status == BGAgentFailed {
-			info.Error = snap.Error
-		}
-		infos = append(infos, info)
-	}
-	return infos, nil
-}
-
-func (q *bgAgentQuerierImpl) TerminateAgent(sessionID, agentID string) error {
-	return q.registry.CancelAgent(sessionID, agentID)
-}
-
-// workshopExecutionBgNotifier implements WorkshopExecutionNotifier by registering
-// workshop step/background executions in bgAgentRegistry so that HasRunningAgents()
-// returns true and the frontend keeps polling for events.
-type workshopExecutionBgNotifier struct {
-	api           *StreamingAPI
-	sessionID     string
-	workspacePath string
-	presetQueryID string
-	userID        string
-}
-
-func (n *workshopExecutionBgNotifier) OnExecutionStart(start todo_creation_human.WorkshopExecutionStart) {
-	if n.api.autoNotificationSessionUnreachable(n.sessionID) {
-		log.Printf("[BG AGENT] OnExecutionStart ignored for stopped session %s (exec=%s)", n.sessionID, start.ID)
-		if start.Cancel != nil {
-			start.Cancel()
-		}
-		return
-	}
-	kind := strings.TrimSpace(start.Kind)
-	if kind == "" {
-		kind = "workshop_background"
-	}
-	if isWorkflowStepTrackingExecution(start.ID, start.Name, start.Metadata) {
-		kind = "workflow_step"
-	}
-	metadata := map[string]string{
-		"workflow_path":    n.workspacePath,
-		"preset_query_id":  n.presetQueryID,
-		"execution_source": trackedExecutionSourceWorkshopBackground,
-	}
-	for k, v := range start.Metadata {
-		if strings.TrimSpace(k) == "" {
-			continue
-		}
-		metadata[k] = v
-	}
-	bgAgent := &BackgroundAgent{
-		ID:                start.ID,
-		ParentExecutionID: start.ParentExecutionID,
-		Name:              start.Name,
-		SessionID:         n.sessionID,
-		Kind:              kind,
-		Status:            BGAgentRunning,
-		CreatedAt:         time.Now(),
-		cancel:            start.Cancel,
-		Metadata:          metadata,
-	}
-	n.api.bgAgentRegistry.Register(n.sessionID, bgAgent)
-	n.api.trackWorkshopExecutionStart(n.sessionID, n.workspacePath, n.presetQueryID, n.userID, start.ID, start.Name)
-
-	// Pre-create the channel so NotifyCompletion never drops a completion
-	n.api.bgAgentRegistry.GetNotificationChannel(n.sessionID)
-
-	// Ensure background completion loop is running
-	n.api.completionLoopStartedMu.Lock()
-	if !n.api.completionLoopStarted[n.sessionID] {
-		n.api.completionLoopStarted[n.sessionID] = true
-		go n.api.backgroundCompletionLoop(n.sessionID)
-	}
-	n.api.completionLoopStartedMu.Unlock()
-
-	// Emit background_agent_started event so BackgroundAgentsStatusBar shows a pill
-	n.api.emitBackgroundAgentEvent(n.sessionID, start.ID, "background_agent_started", map[string]interface{}{
-		"agent_id": start.ID,
-		"name":     start.Name,
-	})
-	n.api.notifyBackgroundAgentStarted(n.sessionID, start.ID)
-}
-
-func isWorkflowStepTrackingExecution(id, name string, meta map[string]string) bool {
-	if meta != nil && strings.TrimSpace(meta["execution_type"]) == "workflow-step" {
-		return true
-	}
-	trimmedName := strings.TrimSpace(name)
-	if strings.HasPrefix(trimmedName, "Step ->") || strings.HasPrefix(trimmedName, "Workflow step ->") {
-		return true
-	}
-	trimmedID := strings.TrimSpace(id)
-	return strings.HasPrefix(trimmedID, "workflow-step-") ||
-		(strings.HasPrefix(trimmedID, "workflow-full-") && strings.Contains(trimmedID, "-step-"))
-}
-
-func (n *workshopExecutionBgNotifier) OnExecutionComplete(execID, name, result string, meta map[string]string, err error) {
-	if n.api.autoNotificationSessionUnreachable(n.sessionID) {
-		n.api.completeTrackedExecution(execID, trackedExecutionStatusCanceled, "session stopped", meta)
-		log.Printf("[BG AGENT] OnExecutionComplete ignored for stopped session %s (exec=%s)", n.sessionID, execID)
-		return
-	}
-	agent := n.api.bgAgentRegistry.Get(n.sessionID, execID)
-	if agent == nil {
-		return
-	}
-
-	// Context-canceled / deadline-exceeded means the execution was cut short
-	// (idle/inactivity timeout, parent-context cancel, etc.) rather than finishing.
-	// Emit a terminated notification so a waiting main agent learns it died — even
-	// when the agent is already marked canceled. This branch runs BEFORE the
-	// already-canceled skip below so a non-user cancel (e.g. a timeout) is not
-	// silently swallowed. MarkTerminalNotified dedups against an explicit stop that
-	// already emitted via OnExecutionTerminated, so we never double-notify.
-	if err != nil && (strings.Contains(err.Error(), "context canceled") || strings.Contains(err.Error(), "context deadline exceeded")) {
-		agent.SetCanceled()
-		n.api.completeTrackedExecution(execID, trackedExecutionStatusCanceled, err.Error(), meta)
-		if agent.MarkTerminalNotified() {
-			log.Printf("[BG AGENT] OnExecutionComplete emitting terminated for context-canceled agent %s", execID)
-			n.api.emitBackgroundAgentEvent(n.sessionID, execID, "background_agent_terminated", map[string]interface{}{
-				"agent_id": execID,
-				"name":     name,
-			})
-			// Drive the auto-notification (synthetic turn / live steer) so the main
-			// agent is actually told, not just the UI pill — matches OnExecutionTerminated.
-			n.api.bgAgentRegistry.NotifyCompletion(n.sessionID, execID)
-		} else {
-			log.Printf("[BG AGENT] OnExecutionComplete skipped duplicate terminated for agent %s", execID)
-		}
-		return
-	}
-
-	// Already canceled with no context-cancel error — an explicit stop
-	// (OnExecutionTerminated / CancelAll) set the flag and already emitted the
-	// terminal event. Don't emit completion events or re-notify.
-	if agent.GetStatus() == BGAgentCanceled {
-		log.Printf("[BG AGENT] OnExecutionComplete skipped for already-canceled agent %s", execID)
-		return
-	}
-
-	duration := time.Since(agent.CreatedAt)
-	if len(meta) > 0 {
-		agent.SetMetadata(meta)
-	}
-	if err != nil {
-		agent.SetError(err.Error())
-		n.api.completeTrackedExecution(execID, trackedExecutionStatusFailed, err.Error(), meta)
-		n.api.emitBackgroundAgentEvent(n.sessionID, execID, "background_agent_completed", map[string]interface{}{
-			"agent_id": execID,
-			"name":     name,
-			"status":   "failed",
-			"error":    err.Error(),
-			"duration": duration.Truncate(time.Second).String(),
-		})
-	} else {
-		agent.SetResult(result) // Store full result — truncation only happens at display/notification time
-		n.api.completeTrackedExecution(execID, trackedExecutionStatusCompleted, "", meta)
-		n.api.emitBackgroundAgentEvent(n.sessionID, execID, "background_agent_completed", map[string]interface{}{
-			"agent_id": execID,
-			"name":     name,
-			"status":   "completed",
-			"result":   truncateForToolResponse(result, 500),
-			"duration": duration.Truncate(time.Second).String(),
-		})
-	}
-
-	// Signal completion to the notification loop (triggers auto-notification synthetic turn).
-	n.api.bgAgentRegistry.NotifyCompletion(n.sessionID, execID)
-}
-
-func (n *workshopExecutionBgNotifier) OnExecutionTerminated(execID, name string) {
-	if n.api.autoNotificationSessionUnreachable(n.sessionID) {
-		n.api.completeTrackedExecution(execID, trackedExecutionStatusCanceled, "session stopped", nil)
-		return
-	}
-	agent := n.api.bgAgentRegistry.Get(n.sessionID, execID)
-	if agent == nil {
-		return
-	}
-	agent.SetCanceled()
-	n.api.completeTrackedExecution(execID, trackedExecutionStatusCanceled, "execution terminated", nil)
-	// Dedup with OnExecutionComplete's context-cancel path: emit the terminated
-	// event only once per agent regardless of which path fires first.
-	if agent.MarkTerminalNotified() {
-		n.api.emitBackgroundAgentEvent(n.sessionID, execID, "background_agent_terminated", map[string]interface{}{
-			"agent_id": execID,
-			"name":     name,
-		})
-	}
-	// Signal completion so the loop can process any pending completions
-	n.api.bgAgentRegistry.NotifyCompletion(n.sessionID, execID)
-}
-
-// workflowSubAgentTrackingNotifier tracks inner workshop sub-agents in the backend
-// execution tree and triggers synthetic-turn notifications only when they finish.
-type workflowSubAgentTrackingNotifier struct {
-	api       *StreamingAPI
-	sessionID string
-}
-
-func (n *workflowSubAgentTrackingNotifier) OnSubAgentStart(start todo_creation_human.WorkshopExecutionStart) {
-	if n == nil || n.api == nil || strings.TrimSpace(start.ID) == "" {
-		return
-	}
-	if n.api.autoNotificationSessionUnreachable(n.sessionID) {
-		if start.Cancel != nil {
-			start.Cancel()
-		}
-		return
-	}
-	kind := strings.TrimSpace(start.Kind)
-	if kind == "" {
-		kind = "workflow_sub_agent"
-	}
-	bgAgent := &BackgroundAgent{
-		ID:                start.ID,
-		ParentExecutionID: start.ParentExecutionID,
-		Name:              start.Name,
-		SessionID:         n.sessionID,
-		Kind:              kind,
-		Status:            BGAgentRunning,
-		CreatedAt:         time.Now(),
-		cancel:            start.Cancel,
-	}
-	n.api.bgAgentRegistry.Register(n.sessionID, bgAgent)
-
-	// Pre-create the completion channel and loop so a fast sub-agent completion
-	// cannot drop its auto-notification. This is only plumbing; no synthetic
-	// turn is emitted until OnSubAgentComplete calls NotifyCompletion.
-	n.api.bgAgentRegistry.GetNotificationChannel(n.sessionID)
-	n.api.completionLoopStartedMu.Lock()
-	if n.api.completionLoopStarted == nil {
-		n.api.completionLoopStarted = make(map[string]bool)
-	}
-	if !n.api.completionLoopStarted[n.sessionID] {
-		n.api.completionLoopStarted[n.sessionID] = true
-		go n.api.backgroundCompletionLoop(n.sessionID)
-	}
-	n.api.completionLoopStartedMu.Unlock()
-
-	n.api.emitBackgroundAgentEvent(n.sessionID, start.ID, "background_agent_started", map[string]interface{}{
-		"agent_id":            start.ID,
-		"name":                start.Name,
-		"parent_execution_id": start.ParentExecutionID,
-	})
-	n.api.notifyBackgroundAgentStarted(n.sessionID, start.ID)
-}
-
-func (n *workflowSubAgentTrackingNotifier) OnSubAgentComplete(agentID, name string, result string, err error) {
-	if n == nil || n.api == nil || strings.TrimSpace(agentID) == "" {
-		return
-	}
-	if n.api.autoNotificationSessionUnreachable(n.sessionID) {
-		return
-	}
-	agent := n.api.bgAgentRegistry.Get(n.sessionID, agentID)
-	if agent == nil {
-		return
-	}
-	if agent.GetStatus() == BGAgentCanceled {
-		return
-	}
-	if err != nil {
-		if strings.Contains(err.Error(), "context canceled") || strings.Contains(err.Error(), "context deadline exceeded") {
-			agent.SetCanceled()
-			// Emit a terminal event for context-canceled sub-agents so the
-			// completion loop has a consistent signal — mirrors OnExecutionComplete
-			// (finding-onsubagentcomplete-context-cancel-silent-drop fix).
-			if !n.api.isSessionMarkedStopped(n.sessionID) {
-				if agent.MarkTerminalNotified() {
-					n.api.emitBackgroundAgentEvent(n.sessionID, agentID, "background_agent_terminated", map[string]interface{}{
-						"agent_id": agentID,
-						"name":     name,
-						"status":   "canceled",
-					})
-					n.api.bgAgentRegistry.NotifyCompletion(n.sessionID, agentID)
-				}
-			}
-			return
-		}
-		agent.SetError(err.Error())
-		if agent.GetStatus() == BGAgentCanceled {
-			return
-		}
-		duration := time.Since(agent.CreatedAt)
-		n.api.emitBackgroundAgentEvent(n.sessionID, agentID, "background_agent_completed", map[string]interface{}{
-			"agent_id": agentID,
-			"name":     name,
-			"status":   "failed",
-			"error":    err.Error(),
-			"duration": duration.Truncate(time.Second).String(),
-		})
-		n.api.bgAgentRegistry.NotifyCompletion(n.sessionID, agentID)
-		return
-	}
-	agent.SetResult(result)
-	if agent.GetStatus() == BGAgentCanceled {
-		return
-	}
-	duration := time.Since(agent.CreatedAt)
-	n.api.emitBackgroundAgentEvent(n.sessionID, agentID, "background_agent_completed", map[string]interface{}{
-		"agent_id": agentID,
-		"name":     name,
-		"status":   "completed",
-		"result":   truncateForToolResponse(result, 500),
-		"duration": duration.Truncate(time.Second).String(),
-	})
-	n.api.bgAgentRegistry.NotifyCompletion(n.sessionID, agentID)
-}
-
-// truncateForToolResponse truncates a string for inclusion in tool responses
-func truncateForToolResponse(s string, maxLen int) string {
-	if len(s) <= maxLen {
-		return s
-	}
-	return s[:maxLen] + "... (truncated)"
-}
-
-// buildCapabilitiesContext creates a CapabilitiesContext from the chat request
-// This is passed to the planner sub-agent so it knows what tools/servers/skills are available
-func buildCapabilitiesContext(req QueryRequest) *virtualtools.CapabilitiesContext {
-	hasBrowser := req.EnableBrowserAccess != nil && *req.EnableBrowserAccess
-	caps := &virtualtools.CapabilitiesContext{
-		EnabledServers: req.EnabledServers,
-		SelectedTools:  req.SelectedTools,
-		HasWorkspace:   true,
-		HasBrowser:     hasBrowser,
-	}
-
-	// Load filesystem skill summaries. Runtime-only skills such as agent-browser
-	// are represented by browser tools and browser prompts instead.
-	workspaceAPIURL := getWorkspaceAPIURL()
-	for _, folderName := range filesystemSelectedSkills(req.SelectedSkills) {
-		skill, err := skills.GetSkill(workspaceAPIURL, folderName)
-		if err != nil {
-			log.Printf("[CAPABILITIES] Warning: Failed to load skill %s: %v", folderName, err)
-			continue
-		}
-		caps.Skills = append(caps.Skills, virtualtools.SkillSummary{
-			Name:        skill.Frontmatter.Name,
-			Description: skill.Frontmatter.Description,
-			FolderName:  folderName,
-		})
-	}
-
-	return caps
-}
-
-// emitDelegationStartEvent emits an event when delegation starts
-// This event serves as the parent for all sub-agent events (via parent_id linking)
-func (api *StreamingAPI) emitDelegationStartEvent(sessionID, delegationID string, depth int, instruction, reasoningLevel, modelID string, servers []string, backgroundAgentID, agentTemplate string) {
-	now := time.Now()
-	eventID := fmt.Sprintf("%s_delegation_start_%s", sessionID, delegationID)
-	eventData := &events.DelegationStartEventData{
-		DelegationID:      delegationID,
-		Depth:             depth,
-		Instruction:       instruction,
-		ReasoningLevel:    reasoningLevel,
-		ModelID:           modelID,
-		Servers:           servers,
-		BackgroundAgentID: backgroundAgentID,
-		AgentTemplate:     agentTemplate,
-		Timestamp:         now.Format(time.RFC3339),
-	}
-	event := events.Event{
-		ID:        eventID,
-		Type:      "delegation_start",
-		Timestamp: now,
-		SessionID: sessionID,
-		Data: &unifiedevents.AgentEvent{
-			Type:           unifiedevents.EventType("delegation_start"),
-			Timestamp:      now,
-			HierarchyLevel: depth,
-			SessionID:      sessionID,
-			Component:      fmt.Sprintf("delegation-%d", depth),
-			CorrelationID:  delegationID, // Links all delegation events together
-			ParentID: func() string {
-				if strings.TrimSpace(backgroundAgentID) == "" {
-					return ""
-				}
-				return fmt.Sprintf("%s_background_agent_started_%s", sessionID, strings.TrimSpace(backgroundAgentID))
-			}(),
-			Data: eventData,
-		},
-	}
-	api.eventStore.AddEvent(sessionID, event)
-	log.Printf("[DELEGATION] Emitted delegation_start event %s for %s at depth %d", eventID, delegationID, depth)
-}
-
-// delegationEndStats holds optional stats for delegation end events
-type delegationEndStats struct {
-	InputTokens  int64
-	OutputTokens int64
-	ToolCalls    int64
-	Duration     string
-	TotalCostUSD float64
-}
-
-// emitDelegationEndEvent emits an event when delegation ends
-// This event has the same correlation_id as delegation_start for grouping
-func (api *StreamingAPI) emitDelegationEndEvent(sessionID, delegationID string, depth int, result, errorMsg string, stats *delegationEndStats) {
-	now := time.Now()
-	delegationStartEventID := fmt.Sprintf("%s_delegation_start_%s", sessionID, delegationID)
-	eventData := &events.DelegationEndEventData{
-		DelegationID: delegationID,
-		Depth:        depth,
-		Result:       result,
-		Error:        errorMsg,
-		Success:      errorMsg == "",
-		Timestamp:    now.Format(time.RFC3339),
-	}
-	if stats != nil {
-		eventData.InputTokens = stats.InputTokens
-		eventData.OutputTokens = stats.OutputTokens
-		eventData.ToolCalls = stats.ToolCalls
-		eventData.Duration = stats.Duration
-		eventData.TotalCostUSD = stats.TotalCostUSD
-	}
-	event := events.Event{
-		ID:        fmt.Sprintf("%s_delegation_end_%s", sessionID, delegationID),
-		Type:      "delegation_end",
-		Timestamp: now,
-		SessionID: sessionID,
-		Data: &unifiedevents.AgentEvent{
-			Type:           unifiedevents.EventType("delegation_end"),
-			Timestamp:      now,
-			HierarchyLevel: depth,
-			SessionID:      sessionID,
-			Component:      fmt.Sprintf("delegation-%d", depth),
-			CorrelationID:  delegationID,           // Links all delegation events together
-			ParentID:       delegationStartEventID, // Makes this a child of delegation_start (for tree display)
-			Data:           eventData,
-		},
-	}
-	api.eventStore.AddEvent(sessionID, event)
-	log.Printf("[DELEGATION] Emitted delegation_end event for %s at depth %d (success: %v)", delegationID, depth, errorMsg == "")
 }
 
 func sanitizeTierModel(model *virtualtools.TierModel) *virtualtools.TierModel {
